// backend/src/types.ts
export interface IUser {
  _id: Types.ObjectId;
  email: string;
  password: string;
  name: string;
  authProvider?: "local" | "google";
  googleId?: string;
  avatar?: string;
  refreshToken?: string;
  apiKey?: string;
  apiKeyExpiresAt?: Date;
  isDeleted: boolean;
  createdAt: Date;
  updatedAt: Date;
}

export interface IUserDocument extends IUser, Document {
  _id: Types.ObjectId;
  isPasswordCorrect(password: string): Promise<boolean>;
  generateAccessToken(): Promise<string>;
  generateRefreshToken(): Promise<string>;
  generateApiKey(expiryDays?: number): string;
}

export interface IUserModel extends Model<IUserDocument> { }

export interface IVisitHistory {
  timestamp: Date;
}

export interface IURL {
  _id: Types.ObjectId;
  userId: Types.ObjectId;
  shortId: string;
  qrCode?: string;
  redirectUrl: string;
  isDeleted: boolean;
  visitHistory: IVisitHistory[];
  createdAt: Date;
  updatedAt: Date;
}
// ...
export interface PaginationInfo {
  currentPage: number;
  totalPages: number;
  totalDocs: number;
  limit?: number;
  hasNextPage: boolean;
  hasPrevPage: boolean;
  nextPage?: number | null;
  prevPage?: number | null;
}

export interface URLWithAnalytics extends IURL {
  totalClicks: number;
  lastVisit?: Date;
  firstVisit?: Date;
}
// ...
export interface GenerateShortURLInput {
  url: string;
  customShortId?: string;
  idLength?: number;
}

export interface GenerateShortURLOutput {
  shortId: string;
  redirectUrl: string;
  fullShortUrl: string;
  isExisting?: boolean;
  isCustom?: boolean;
}

export interface UpdateShortURLInput {
  shortId: string;
  qrCode: string;
}

export interface UpdateShortURLOutput {
  shortId: string;
  redirectUrl: string;
  qrCode: string;
  fullShortUrl: string;
  updatedAt: Date;
}

export interface AnalyticsOutput {
  totalClicks: number;
  analytics: IVisitHistory[];
}

export interface DeleteURLOutput {
  shortId: string;
  isDeleted: boolean;
}